requirements/base.txt
# Core dependencies
python-dotenv==1.0.1
pydantic==2.6.3
psycopg2-binary==2.9.9
confluent-kafka==2.9.0
redis==5.0.1
fastapi==0.110.0
uvicorn[standard]==0.27.1
requests==2.31.0

# Monitoring and logging
prometheus-client==0.20.0
structlog==24.1.0
python-json-logger==2.0.7

# Utilities
tenacity==8.2.3
httpx==0.27.0
pytz==2024.1 